Fixed crate dependency issue
The module text_input was loaded at line 80, which conflicts with the crate text_input as declared on line 15. Cargo suggests binding the crate name as other_text_input, however, I named it crate_text_input to be more descriptive. I do not know if it will be an issue later in development, but libcosmic does at least compile successfully.