applet: Add `icon_button_from_handle` (#368)
`StatusNotifierItem` can use either named icons, or pixmap data. So to
support the latter, it needs a way to show a non-named icon with the
same button styling as any other planel applet icon.

The `StatusNotifierItem` protocol actually supports sending different
icons with different sizes, so ideally we'd want a `icon::Data` variant
that contains multiple sizes? But just using the largest and scaling
works for now.