Call systemd `StartTransientUnit` what starting app
This is needed for things like `xdg-desktop-portal` to get the app ID
from a pid, in unsandboxed apps.

https://systemd.io/DESKTOP_ENVIRONMENTS documents this, and this sets
things similarly to how Gnome does. Which should be good for now.

Making `spawn_desktop_exec` an `async` function is reasonable given it
is called in `async` functions in
`cosmic-app-list`/`cosmic-launcher`/`cosmic-applibrary`.