Fix image-button example
The image-button example did not work with `just run image-button`.
The first problem was that the package name was `cosmic-image-button`
and not `image-button`. It was defined as `image-button` in the
justfile.

The second problem was that the `wayland` feature defines the
`set_window_title` as taking an `Id`. The easiest solution was to just
use the `winit` feature instead. We can alternatively pass
`window::Id::MAIN` if needed.