Add `dbus_connection()` method to `app::Application` trait
This allows an application that wants to expose a DBus protocol to use
the same session DBus connection, and provide a protocol on the same
owned bus name.

Without this, a Cosmic app using `single-instance` that wants to provide
a DBus protocol must provide it on a bus name other than `APP_ID`.