fix(cosmic_config): Make `get_local` return `key_path()` error
It looks like this was changed to be like this in cd8f4ee85. But
`key_path()` only seems to return `Error::NoConfigDirectory` or
`Error::InvalidName`, so it seems reasonable to just update `get()` to
also check for `NoConfigDirectory`. (`get_local()` doesn't seem
to be used elsewhere, so I don't see nay issues this can cause.)